refactor(errors): #301 promote ErrParquetSetInconsistent to the root package
internal/httpapi must classify it to redact its S3 object keys from public 5xx
bodies, and cannot import internal/federated without pulling DuckDB CGO into a
pure-Go test build. Aliased back per the #299 precedent, so every existing call
site and test compiles unchanged.
